+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+- Output the original image when `canvas.getContext('2d').getImageData()` is unavailable in Firefox's `privacy.resistFingerprinting` mode (#177).
+
 ## 1.2.1 (Feb 28, 2023)
 
 - Fix incompatible syntax in the bundled files (#170).